+++ crates/sui-json-rpc/src/read_api.rs
@@ -8,7 +8,6 @@ use move_binary_format::normalized::{Module as NormalizedModule, Type};
 use move_core_types::identifier::Identifier;
 use std::collections::BTreeMap;
 use std::sync::Arc;
-use sui_types::committee::EpochId;
 use sui_types::intent::{AppId, Intent, IntentMessage, IntentScope, IntentVersion};
 use tap::TapFallible;
 
@@ -17,16 +16,16 @@ use fastcrypto::encoding::Base64;
 use jsonrpsee::RpcModule;
 use sui_core::authority::AuthorityState;
 use sui_json_rpc_types::{
-    Checkpoint, CheckpointId, DevInspectResults, DynamicFieldPage, GetObjectDataResponse,
-    GetPastObjectDataResponse, GetRawObjectDataResponse, MoveFunctionArgType, ObjectValueKind,
-    Page, SuiMoveNormalizedFunction, SuiMoveNormalizedModule, SuiMoveNormalizedStruct,
-    SuiObjectInfo, SuiTransactionEffects, SuiTransactionResponse, TransactionsPage,
+    Checkpoint, CheckpointId, DynamicFieldPage, GetObjectDataResponse, GetPastObjectDataResponse,
+    GetRawObjectDataResponse, MoveFunctionArgType, ObjectValueKind, Page,
+    SuiMoveNormalizedFunction, SuiMoveNormalizedModule, SuiMoveNormalizedStruct, SuiObjectInfo,
+    SuiTransactionEffects, SuiTransactionResponse, TransactionsPage,
 };
 use sui_open_rpc::Module;
 use sui_types::base_types::SequenceNumber;
 use sui_types::base_types::{ObjectID, SuiAddress, TransactionDigest, TxSequenceNumber};
 use sui_types::crypto::sha3_hash;
-use sui_types::messages::{TransactionData, TransactionKind};
+use sui_types::messages::TransactionData;
 use sui_types::messages_checkpoint::{
     CheckpointContents, CheckpointContentsDigest, CheckpointDigest, CheckpointSequenceNumber,
     CheckpointSummary,
@@ -51,14 +50,6 @@ pub struct ReadApi {
 impl ReadApi {
     pub fn new(state: Arc<AuthorityState>) -> Self {
         Self { state }
-    }
-
-    fn get_sui_system_state_object_epoch_and_gas_price(&self) -> RpcResult<(EpochId, u64)> {
-        let sys_state = self
-            .state
-            .get_sui_system_state_object()
-            .map_err(|e| anyhow!("Unable to retrieve sui system state object: {e}"))?;
-        Ok((sys_state.epoch, sys_state.reference_gas_price))
     }
 
     fn get_checkpoint_internal(&self, id: CheckpointId) -> Result<Checkpoint, Error> {
@@ -173,40 +164,6 @@ impl ReadApiServer for ReadApi {
             confirmed_local_execution: None,
             checkpoint: checkpoint.map(|(_epoch, checkpoint)| checkpoint),
         })
-    }
-
-    async fn dev_inspect_transaction(
-        &self,
-        sender_address: SuiAddress,
-        tx_bytes: Base64,
-        gas_price: Option<u64>,
-        epoch: Option<EpochId>,
-    ) -> RpcResult<DevInspectResults> {
-        let (mut current_epoch, mut reference_gas_price) = (0, 0);
-        // Only fetch from DB if necessary
-        if gas_price.is_none() || epoch.is_none() {
-            (current_epoch, reference_gas_price) =
-                self.get_sui_system_state_object_epoch_and_gas_price()?
-        }
-        let tx_kind: TransactionKind =
-            bcs::from_bytes(&tx_bytes.to_vec().map_err(|e| anyhow!(e))?).map_err(|e| anyhow!(e))?;
-        Ok(self
-            .state
-            .dev_inspect_transaction(
-                sender_address,
-                tx_kind,
-                gas_price.unwrap_or(reference_gas_price),
-                epoch.unwrap_or(current_epoch),
-            )
-            .await?)
-    }
-
-    async fn dry_run_transaction(&self, tx_bytes: Base64) -> RpcResult<SuiTransactionEffects> {
-        let (txn_data, txn_digest) = get_transaction_data_and_digest(tx_bytes)?;
-        Ok(self
-            .state
-            .dry_exec_transaction(txn_data, txn_digest)
-            .await?)
     }
 
     async fn get_normalized_move_modules_by_package(
