+++ crates/sui-json-rpc/src/transaction_execution_api.rs
@@ -1,7 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::api::TransactionExecutionServer;
+use crate::api::WriteApiServer;
+use crate::read_api::get_transaction_data_and_digest;
 use crate::SuiRpcModule;
 use anyhow::anyhow;
 use async_trait::async_trait;
@@ -12,35 +13,49 @@ use jsonrpsee::RpcModule;
 use move_bytecode_utils::module_cache::SyncModuleCache;
 use mysten_metrics::spawn_monitored_task;
 use std::sync::Arc;
-use sui_core::authority::{AuthorityStore, ResolverWrapper};
+use sui_core::authority::{AuthorityState, AuthorityStore, ResolverWrapper};
 use sui_core::authority_client::NetworkAuthorityClient;
 use sui_core::transaction_orchestrator::TransactiondOrchestrator;
-use sui_json_rpc_types::{SuiTransactionEffects, SuiTransactionResponse};
+use sui_json_rpc_types::{DevInspectResults, SuiTransactionEffects, SuiTransactionResponse};
 use sui_open_rpc::Module;
+use sui_types::base_types::{EpochId, SuiAddress};
 use sui_types::intent::Intent;
-use sui_types::messages::{ExecuteTransactionRequest, ExecuteTransactionRequestType};
+use sui_types::messages::{
+    ExecuteTransactionRequest, ExecuteTransactionRequestType, TransactionKind,
+};
 use sui_types::messages::{ExecuteTransactionResponse, Transaction};
 use sui_types::signature::GenericSignature;
 
 pub struct TransactionExecutionApi {
-    pub transaction_orchestrator: Arc<TransactiondOrchestrator<NetworkAuthorityClient>>,
-    pub module_cache: Arc<SyncModuleCache<ResolverWrapper<AuthorityStore>>>,
+    state: Arc<AuthorityState>,
+    transaction_orchestrator: Arc<TransactiondOrchestrator<NetworkAuthorityClient>>,
+    module_cache: Arc<SyncModuleCache<ResolverWrapper<AuthorityStore>>>,
 }
 
 impl TransactionExecutionApi {
     pub fn new(
+        state: Arc<AuthorityState>,
         transaction_orchestrator: Arc<TransactiondOrchestrator<NetworkAuthorityClient>>,
         module_cache: Arc<SyncModuleCache<ResolverWrapper<AuthorityStore>>>,
     ) -> Self {
         Self {
+            state,
             transaction_orchestrator,
             module_cache,
         }
     }
+
+    fn get_sui_system_state_object_epoch_and_gas_price(&self) -> RpcResult<(EpochId, u64)> {
+        let sys_state = self
+            .state
+            .get_sui_system_state_object()
+            .map_err(|e| anyhow!("Unable to retrieve sui system state object: {e}"))?;
+        Ok((sys_state.epoch, sys_state.reference_gas_price))
+    }
 }
 
 #[async_trait]
-impl TransactionExecutionServer for TransactionExecutionApi {
+impl WriteApiServer for TransactionExecutionApi {
     async fn execute_transaction(
         &self,
         tx_bytes: Base64,
@@ -109,6 +124,40 @@ impl TransactionExecutionServer for TransactionExecutionApi {
             }
         }
     }
+
+    async fn dev_inspect_transaction(
+        &self,
+        sender_address: SuiAddress,
+        tx_bytes: Base64,
+        gas_price: Option<u64>,
+        epoch: Option<EpochId>,
+    ) -> RpcResult<DevInspectResults> {
+        let (mut current_epoch, mut reference_gas_price) = (0, 0);
+        // Only fetch from DB if necessary
+        if gas_price.is_none() || epoch.is_none() {
+            (current_epoch, reference_gas_price) =
+                self.get_sui_system_state_object_epoch_and_gas_price()?
+        }
+        let tx_kind: TransactionKind =
+            bcs::from_bytes(&tx_bytes.to_vec().map_err(|e| anyhow!(e))?).map_err(|e| anyhow!(e))?;
+        Ok(self
+            .state
+            .dev_inspect_transaction(
+                sender_address,
+                tx_kind,
+                gas_price.unwrap_or(reference_gas_price),
+                epoch.unwrap_or(current_epoch),
+            )
+            .await?)
+    }
+
+    async fn dry_run_transaction(&self, tx_bytes: Base64) -> RpcResult<SuiTransactionEffects> {
+        let (txn_data, txn_digest) = get_transaction_data_and_digest(tx_bytes)?;
+        Ok(self
+            .state
+            .dry_exec_transaction(txn_data, txn_digest)
+            .await?)
+    }
 }
 
 impl SuiRpcModule for TransactionExecutionApi {
@@ -117,6 +166,6 @@ impl SuiRpcModule for TransactionExecutionApi {
     }
 
     fn rpc_doc_module() -> Module {
-        crate::api::TransactionExecutionOpenRpc::module_doc()
+        crate::api::WriteApiOpenRpc::module_doc()
     }
 }
